Change some messages to debug only
When a source location can't be found, we now use debug(), so only that we only spam people who are in debug mode. This means if a basenot source name is not found in the DB, it will simply not give any information. Same goes for if it was unable to calculate a destination for the lost contact. It silently fails. Requested by Spro.